IsoBank-README.md
IsoBank is a data repository for stable isotope ecologists. It contains stable isotope rations of plants, animals and inorganic materials along with relevant metadata.

This GItHub page is a focal point for a collaboration of ecologists and data scientists working to create, refine and apply IsoBank to further ecological knowledge.